requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.3.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==2.2
djangorestframework==3.12.4
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==4.0.1
jdcal==1.4.1
joblib==1.0.1
kiwisolver==1.3.1
kombu==4.6.11
matplotlib==2.2.3
numpy==1.15.4
openpyxl==2.5.6
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.1.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.4.1
